split into subcommand for prepare and full test
