+++ Run-AndroidTests.ps1
@@ -1,1 +1,1 @@
-dotnet run --project src/Xappium.Cli/Xappium.Cli.csproj -- -uitest sample/TestApp.UITests/TestApp.UITests.csproj -app sample/TestApp.Android/TestApp.Android.csproj
+dotnet run --project src/Xappium.Cli/Xappium.Cli.csproj -- test -uitest sample/TestApp.UITests/TestApp.UITests.csproj -app sample/TestApp.Android/TestApp.Android.csproj
