+++ src/Xappium.Cli/Utilities/DelegateDisposable.cs
@@ -1,0 +1,23 @@
+﻿using System;
+
+namespace Xappium.Utilities
+{
+    internal class DelegateDisposable : IDisposable
+    {
+        private WeakReference<Action> _weakDelegate;
+
+        public DelegateDisposable(Action @delegate)
+        {
+            _weakDelegate = new WeakReference<Action>(@delegate);
+        }
+
+        public void Dispose()
+        {
+            if(_weakDelegate != null && _weakDelegate.TryGetTarget(out var target))
+            {
+                target?.Invoke();
+                _weakDelegate = null;
+            }
+        }
+    }
+}
